+++ test/unit/Chain.js
@@ -1,7 +1,24 @@
-import { Chain, Accumulator, Muxer, Demuxer, Mapper } from '../../src/graflow'
+import {Chain, Mapper} from '../../src/graflow'
 
 describe('Chain', () => {
   it('should return a component that joins components together', () => {
+    const comp = Chain(
+      Mapper(v => v * 2),
+      Mapper(v => v + 1)
+    )
+
+    const listener = spy()
+
+    comp.on(listener)
+
+    comp.send(5)
+
+    expect(listener).to.have.been.calledOnce
+    expect(listener.getCall(0).args[0]).to.be.equal(11)
+  })
+
+/*
+it('should return a component that joins components together', () => {
     const comp = Chain(
       Muxer('dividend', 'divisor')
       , Accumulator()
@@ -15,12 +32,12 @@ describe('Chain', () => {
     const listenerQuotient = spy()
     const listenerRemainder = spy()
 
-    comp.out.quotient.on(listenerQuotient)
-    comp.out.remainder.on(listenerRemainder)
+    comp.on('quotient', listenerQuotient)
+    comp.on('remainder', listenerRemainder)
 
     // 26 / 11 = 2 remainder 4
-    comp.in.dividend.send(26)
-    comp.in.divisor.send(11)
+    comp.send('dividend', 26)
+    comp.send('divisor', 11)
 
     expect(listenerQuotient).to.have.been.calledOnce
     expect(listenerQuotient.getCall(0).args[0]).to.be.equal(2)
@@ -28,4 +45,5 @@ describe('Chain', () => {
     expect(listenerRemainder).to.have.been.calledOnce
     expect(listenerRemainder.getCall(0).args[0]).to.be.equal(4)
   })
+*/
 })