+++ test/unit/UniqueFilter.js
@@ -5,17 +5,19 @@ describe('UniqueFilter', () => {
     const norepeater = UniqueFilter()
     const listener = spy()
 
-    norepeater.out.default.on(listener)
+    norepeater.on(listener)
 
-    norepeater.in.default.send(1)
+    norepeater.send(1)
 
     expect(listener).to.have.been.calledOnce
     expect(listener.getCall(0).args[0]).to.be.equal(1)
 
-    norepeater.in.default.send(1)
+    norepeater.send(1)
+
     expect(listener).to.have.been.calledOnce
 
-    norepeater.in.default.send(2)
+    norepeater.send(2)
+
     expect(listener).to.have.been.calledTwice
     expect(listener.getCall(1).args[0]).to.be.equal(2)
   })